Move find_symbol out of TraceCompiler.
It doesn't really have anything to do with a Trace Compiler.

Requires moving some other bits around too.
